Generate the attr and schema classes using a naive code generator
The data objects are created using attrs and use type annotation which
improves code completion in editors. The schema classes use marshmallow.
The service files are not generated yet, this might be something for the
future.

The code generator is still really hacky and is definately not the
most clean code :-) It should be improved in the future. It currently
expects the commercetools raml file to be located at
../commercetools-api-reference/api/raml
See https://github.com/commercetools/commercetools-api-reference

Currently only the product create and update are implemented.