Remove default values for required values in type constructors
Previously all properties were marked as either optional or required
based on the raml specification. The __init__ method of the types were
however always optional since the default value of None was defined.

This commit re-orders the keyword arguments and removes default values
for required values. This means that all values which are required
should be specified when creating the instance.

Note that this is a backwards incompatible change